+++ src/python/autotransform/event/util.py
@@ -1,0 +1,57 @@
+# AutoTransform
+# Large scale, component based code modification library
+#
+# Licensed under the MIT License <http://opensource.org/licenses/MIT>
+# SPDX-License-Identifier: MIT
+# Copyright (c) 2022-present Nathan Rockenbach <http://github.com/nathro>
+
+# @black_format
+
+"""The Util events are used to handle events related to AutoTransform utilities.
+"""
+
+from typing import TypedDict
+
+from autotransform.event.base import Event
+from autotransform.event.logginglevel import LoggingLevel
+from autotransform.event.type import EventType
+
+
+class RevertFileEventData(TypedDict):
+    """The data for a RevertFileEvent. Contains the information that will be
+    logged when the event is triggered."""
+
+    file_path: str
+
+
+class RevertFileEvent(Event[RevertFileEventData]):
+    """A simple, generic event used to log file revert information to the console."""
+
+    @staticmethod
+    def get_type() -> EventType:
+        """Used to represent the type of Event, output to logs.
+
+        Returns:
+            EventType: The unique type associated with this Event.
+        """
+
+        return EventType.REVERT_FILE
+
+    @staticmethod
+    def get_logging_level() -> LoggingLevel:
+        """The logging level for events of this type.
+
+        Returns:
+            LoggingLevel: The logging detail required to log this event.
+        """
+
+        return LoggingLevel.VERBOSE
+
+    def _get_message(self) -> str:
+        """Gets a message representing the details of the event.
+
+        Returns:
+            str: The message for the event.
+        """
+
+        return self.data["file_path"]